+++ index.html
@@ -7,8 +7,10 @@
     <link rel="stylesheet" href="./index.css">
 </head>
 <body> 
+    <h1>Awesome books</h1>
+    <section class="dateTime" id="displayDate">
+    </section>
     <section id="list_section">
-        <h1>Awesome books</h1>
         <div id="list" class="list">
 
         </div>
