+++ modules/create.js
@@ -1,0 +1,6 @@
+export const create = () => {
+  if (localStorage.getItem('books') == null) {
+    localStorage.setItem('books', JSON.stringify([]));
+  }
+  return JSON.parse(localStorage.getItem('books'));
+};
